% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{az_login}
\alias{az_login}
\title{Log in to Azure}
\usage{
az_login(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=az_cli_run]{az_cli_run}}
  \describe{
    \item{\code{cmd}}{Command, required, optional, additional parameters, as for \code{\link[processx:run]{processx::run()}}
\code{add} parameters are reserved for the user to pass down additional arguments.}
    \item{\code{req}}{Command, required, optional, additional parameters, as for \code{\link[processx:run]{processx::run()}}
\code{add} parameters are reserved for the user to pass down additional arguments.}
    \item{\code{opt}}{Command, required, optional, additional parameters, as for \code{\link[processx:run]{processx::run()}}
\code{add} parameters are reserved for the user to pass down additional arguments.}
    \item{\code{add}}{Command, required, optional, additional parameters, as for \code{\link[processx:run]{processx::run()}}
\code{add} parameters are reserved for the user to pass down additional arguments.}
    \item{\code{echo_cmd}}{Whether to print the command to run to the screen.}
    \item{\code{echo}}{Whether to print the standard output and error
to the screen. Note that the order of the standard output and error
lines are not necessarily correct, as standard output is typically
buffered.}
  }}
}
\description{
Helper for interactive login.
Do not script this function unless you know what you are doing.
}
\seealso{
Other azure functions: 
\code{\link{az_account}()},
\code{\link{az_cli_run}()},
\code{\link{az_configure}()},
\code{\link{az_webapp}()},
\code{\link{include_app2_az}()},
\code{\link{shiny_deploy_az}()},
\code{\link{shiny_opts_az}()}
}
\concept{azure functions}
