% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{az_webapp_shiny_opts}
\alias{az_webapp_shiny_opts}
\title{Shiny options for Azure}
\usage{
az_webapp_shiny_opts()
}
\description{
Set shiny options as \href{https://docs.microsoft.com/en-us/azure/app-service/containers/configure-custom-container}{required for an Azure Webapp}:
\itemize{
\item \verb{options(shiny.port = as.integer(Sys.getenv('PORT'))}.
Your custom container is expected to listen on \code{PORT}, an environment variable set by Azure.
If your image suggests \code{EXPOSE}d ports, that may be respected by Azure (undocumented behavior).
\item \code{options(shiny.host = "0.0.0.0")} to make your shiny application accessable to the Azure Webapp hosting environment.
}
}
\details{
You can also set these options manually as in the below example.
}
\seealso{
Other azure functions: 
\code{\link{az_webapp_show}()},
\code{\link{az_webapp}()},
\code{\link{include_app2_azure}()}
}
\concept{azure functions}
