% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{az_webapp_show}
\alias{az_webapp_show}
\title{Get details of a web app}
\usage{
az_webapp_show(name, resource_group, slot = NULL, subscription, ...)
}
\description{
Get details of a web app
}
\seealso{
Other azure functions: 
\code{\link{az_webapp_shiny_opts}()},
\code{\link{az_webapp}()},
\code{\link{include_app2_azure}()}
}
\concept{azure functions}
