% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{buildtime_sha}
\alias{buildtime_sha}
\title{SHA at build time}
\format{
An object of class \code{character} of length 1.
}
\usage{
buildtime_sha
}
\description{
Used for testing.
Because package is installed with remotes::install_local() on GitHub Actions, no other source of the commit at \emph{runtime} inside
}
\keyword{internal}
