% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_tag}
\alias{get_tag}
\title{Use \code{github.sha} environment variable or \code{"latest"} as docker image tag}
\usage{
get_tag()
}
\value{
character string
}
\description{
Retrieves the current git commit SHA from the \href{https://docs.github.com/en/actions/reference/context-and-expression-syntax-for-github-actions}{\code{github.sha}} environment variable set on GitHub Actions.
Returns \code{latest}, if the environment variable is an empty string.
Internal function used to retrieve the appropriate docker image tag for a shiny app.
}
\keyword{internal}
