% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getenv2}
\alias{getenv2}
\title{Get GitHub Actions Environment Variables or Local Equivalent}
\usage{
getenv2(ghactions, local)
}
\arguments{
\item{ghactions}{The \href{https://docs.github.com/en/actions/configuring-and-managing-workflows/using-environment-variables}{GitHub Actions environment variable} to get when \code{is_github_actions() == TRUE}, as a character string.}

\item{local}{The value to be used when \code{is_github_actions() == FALSE}.}
}
\description{
Using \href{https://docs.github.com/en/actions/configuring-and-managing-workflows/using-environment-variables}{environment variables set on GitHub Actions} \emph{or} their local equivalents can sometimes be useful, for example in testing.
When running inside GitHub Actions (\code{\link[=is_github_actions]{is_github_actions()}}), returns the respective environment variable.
Otherwise, returns a \code{local} equivalent or errors out.
}
\keyword{internal}
