% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{include_app2}
\alias{include_app2}
\title{Embed a shiny app in a pkgdown site}
\usage{
include_app2(src, name = NULL, height = "100\%", width = "100\%")
}
\arguments{
\item{src}{The URL of the shiny app to embed.}

\item{name}{Name of the iframe.}

\item{height, width}{The dimensions of the iframe as CSS units.}
}
\value{
A tag list as from \code{\link[htmltools:tag]{htmltools::tagList()}}.
}
\description{
Embeds a shiny app in a pkgdown, or similar bootstrap website.
Does not offer a print method for non-html output.
}
\concept{helper functions}
