% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{include_app2_azure}
\alias{include_app2_azure}
\title{Embed a shiny app with Microsoft Azure Web Apps for Containers}
\usage{
include_app2_azure(slot = NULL, ...)
}
\arguments{
\item{slot}{The name of the \href{https://docs.microsoft.com/en-us/azure/app-service/deploy-staging-slots}{deployment slot}.
Defaults to the production slot if not specified.
Only available for higher app service plan tiers.}

\item{...}{
  Arguments passed on to \code{\link[=include_app2]{include_app2}}
  \describe{
    \item{\code{src}}{The URL of the shiny app to embed.}
    \item{\code{name}}{Name of the iframe.}
    \item{\code{height}}{The dimensions of the iframe as CSS units.}
    \item{\code{width}}{The dimensions of the iframe as CSS units.}
  }}
}
\description{
Embed a shiny app with Microsoft Azure Web Apps for Containers
}
\seealso{
Other azure functions: 
\code{\link{az_webapp_shiny_opts}()},
\code{\link{az_webapp_show}()},
\code{\link{az_webapp}()}
}
\concept{azure functions}
