% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{shiny_deploy_az}
\alias{shiny_deploy_az}
\title{Deploy a shiny app to Azure}
\usage{
shiny_deploy_az(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=az_account]{az_account}}, \code{\link[=az_configure]{az_configure}}, \code{\link[=az_webapp]{az_webapp}}, \code{\link[=az_webapp_create]{az_webapp_create}}, \code{\link[=az_webapp_config_container_set]{az_webapp_config_container_set}}, \code{\link[=az_cli_run]{az_cli_run}}
  \describe{
    \item{\code{subscription}}{Name or ID of the Azure subscription to which costs are billed.
According to an upvoted answer on Stack Overflow, \href{https://stackoverflow.com/questions/45661109/are-azure-subscription-id-aad-tenant-id-and-aad-app-client-id-considered-secre}{Azure subscription IDs need not be considered a secret or personal identifiable information (PII)}.
However, depending your applicable context and policies, you may want to provide this argument as a secret.

To find out which subscriptions you are currently authorised to use, run \code{print(az_account_list())}.}
    \item{\code{name}}{Name of the web app.
(In the Azure CLI, this argument is sometimes known as \code{name}, and sometimes as \code{web}).}
    \item{\code{resource_group}}{The \href{https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/manage-resource-groups-portal}{Azure resource group} to which the shiny app should belong.}
    \item{\code{restart}}{whether to restart the web app.}
    \item{\code{slot}}{The name of the \href{https://docs.microsoft.com/en-us/azure/app-service/deploy-staging-slots}{deployment slot}.
Defaults to the production slot if not specified.
Only available for higher app service plan tiers.}
    \item{\code{plan}}{Name or resource id of the app service plan.}
    \item{\code{startup_file}}{\verb{docker run} \href{https://docs.docker.com/engine/reference/run/}{\verb{[COMMAND]}} to use inside of your custom image \code{deployment_container_image_name}.
Defaults to \code{NULL}, in which case the container is expected to start up shiny automatically (recommended).
For details on the shiny startup command, see the examples.

\strong{The \verb{[EXPR]} (anything after \code{-e}) must not be quoted, and must not contain spaces (\href{https://github.com/subugoe/shinycaas/issues/27}{#27})}.
For example, the following \code{startup-file}s are valid (if nonsensical, because they don't start a shiny app)
\itemize{
\item \code{"Rscript -e 1+1"} (no spaces)
\item \code{"Rscript -e print('foo')"} (no spaces, no quoting \emph{of} the \verb{[EXPR]})
}

The following \code{startup-file}s are \emph{invalid}:
\itemize{
\item \code{"Rscript -e 1 + 1"} (spaces inside \verb{[EXPR]})
\item \code{"Rscript -e '1+1'"} (quoting of \verb{[EXPR]} would be treated as \verb{"Rscript -e '\\"1+1\\"'"}).
}}
    \item{\code{deployment_container_image_name}}{The custom image name and optionally the tag name.
Image must
\itemize{
\item include everything needed to run the shiny app, including shiny itself,
but \emph{does not} need to include shiny server or other software to route, load balance and serve shiny,
\item include an \code{ENTRYPOINT} and/or \href{https://docs.docker.com/engine/reference/builder/#cmd}{\code{CMD}} instruction to start shiny automatically (recommended), \emph{or} shiny must be started via the \code{startup_file} argument.
}}
    \item{\code{docker_registry_server_url}}{The container registry server url.
Defaults to \code{NULL}, in which case the azure default, \href{http://hub.docker.com}{docker hub} is used.}
    \item{\code{docker_registry_server_user}}{Credentials for private container registries.
Defaults to \code{NULL} for public registries.
Do not expose your credentials in public code; it's best to use secret environment variables.}
    \item{\code{docker_registry_server_password}}{Credentials for private container registries.
Defaults to \code{NULL} for public registries.
Do not expose your credentials in public code; it's best to use secret environment variables.}
    \item{\code{cmd}}{Command, required, optional, additional parameters, as for \code{\link[processx:run]{processx::run()}}
\code{add} parameters are reserved for the user to pass down additional arguments.}
    \item{\code{req}}{Command, required, optional, additional parameters, as for \code{\link[processx:run]{processx::run()}}
\code{add} parameters are reserved for the user to pass down additional arguments.}
    \item{\code{opt}}{Command, required, optional, additional parameters, as for \code{\link[processx:run]{processx::run()}}
\code{add} parameters are reserved for the user to pass down additional arguments.}
    \item{\code{add}}{Command, required, optional, additional parameters, as for \code{\link[processx:run]{processx::run()}}
\code{add} parameters are reserved for the user to pass down additional arguments.}
    \item{\code{echo_cmd}}{Whether to print the command to run to the screen.}
    \item{\code{echo}}{Whether to print the standard output and error
to the screen. Note that the order of the standard output and error
lines are not necessarily correct, as standard output is typically
buffered.}
  }}
}
\description{
Calls the  with defaults suitable for deploying a shiny app.
Wraps several steps.
}
\examples{
# hoad credentials used for testing
# authentication happens outside of r, see github actions main.yaml
# during testing .azure/ is not available, so we have to set that here
plan <- "hoad"
resource_group <- "hoad"

# deploy shiny app using rocker image
shiny_deploy_az(
  name = "hello-shiny",
  # this image actually includes *more* than necessary
  # for example, it includes shinyserver, but just shiny would suffice
  deployment_container_image_name = "rocker/shiny:4.0.2",
  # above image has no `ENTRYPOINT` and/or `CMD` to start shiny by default.
  # so this `[COMMAND]` must be appended to `docker run`
  startup_file = paste(
    "Rscript",
    # setting shiny options for azure manually
    # equivalent to running shinycaas::shiny_opts_az()
    "-e options(shiny.host='0.0.0.0',shiny.port=as.integer(Sys.getenv('PORT')))",
    # remove getOption call https://github.com/subugoe/shinycaas/issues/37
    "-e shiny::runExample('01_hello',port=getOption('shiny.port'))"
  ),
  plan = plan,
  resource_group = resource_group
)

# deploy shiny app to slot
shiny_deploy_az(
  name = "hello-shiny",
  deployment_container_image_name = "rocker/shiny:4.0.2",
  startup_file = paste(
    "Rscript",
    "-e options(shiny.host='0.0.0.0',shiny.port=as.integer(Sys.getenv('PORT')))",
    "-e shiny::runExample('04_mpg',port=getOption('shiny.port'))"
  ),
  plan = plan,
  slot = "mpg" # a more suitable slot name might be "dev" or "staging"
)

# below env vars and secrets are only available on github actions
if (is_github_actions()) {
  # deploy shiny app using muggle image
  # for an easier way to set these arguments, see the {muggle} package
  shiny_deploy_az(
    name = "old-faithful",
    deployment_container_image_name = paste0(
      "docker.pkg.github.com/subugoe/shinycaas/oldfaithful", ":",
      ifelse(is_github_actions(), Sys.getenv("GITHUB_SHA"), "latest")
    ),
    plan = plan,
    docker_registry_server_url = "https://docker.pkg.github.com",
    docker_registry_server_user = Sys.getenv("GITHUB_ACTOR"),
    docker_registry_server_password = Sys.getenv("GITHUB_TOKEN")
  )
}
}
\seealso{
Other azure functions: 
\code{\link{az_account}()},
\code{\link{az_cli_run}()},
\code{\link{az_configure}()},
\code{\link{az_login}()},
\code{\link{az_webapp}()},
\code{\link{include_app2_az}()},
\code{\link{shiny_opts_az}()}
}
\concept{azure functions}
