% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinycaas-package.R
\docType{package}
\name{shinycaas-package}
\alias{shinycaas}
\alias{shinycaas-package}
\title{shinycaas: Deploy shiny applications to container as a service (CaaS)}
\description{
Deploying shiny apps to container as a service (CaaS) products on public cloud vendors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://subugoe.github.io/shinycaas}
  \item \url{https://github.com/subugoe/shinycaas}
  \item Report bugs at \url{https://github.com/subugoe/shinycaas/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Held \email{info@maxheld.de} (\href{https://orcid.org/0000-0002-4703-5388}{ORCID})

Other contributors:
\itemize{
  \item Göttingen State and University Library [copyright holder, funder]
}

}
\keyword{internal}
