% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{az_webapp_config}
\alias{az_webapp_config}
\title{Run a shiny app on \href{https://azure.microsoft.com/en-us/services/app-service/containers/}{Microsoft Azure Web Apps for Containers} (awac)}
\usage{
az_webapp_config(
  name = "hoad",
  docker_custom_image_name = paste0("docker.pkg.github.com/subugoe/hoad/hoad-dev", ":",
    "azure"),
  startup_file = "init.sh",
  subscription = "f0dd3a37-0a4e-4e7f-9c9b-cb9f60146edc",
  resource_group = "hoad",
  docker_registry_server_url = "https://docker.pkg.github.com",
  docker_registry_server_user = "maxheld83"
)
}
\arguments{
\item{name}{Name of the web app.}

\item{docker_custom_image_name}{The container custom image name and optionally the tag name.
Must include everything to run the shiny app, including shiny itself.
Does not need to include shiny server or other software to route, load balance and serve shiny.}

\item{startup_file}{Command use as an \href{https://docs.docker.com/engine/reference/run/#entrypoint-default-command-to-execute-at-runtime}{\code{--entrypoint}}.}

\item{subscription}{Name or ID of the Azure subscription to which costs are billed.
According to an upvoted answer on Stack Overflow, \href{https://stackoverflow.com/questions/45661109/are-azure-subscription-id-aad-tenant-id-and-aad-app-client-id-considered-secre}{Azure subscription IDs need not be considered a secret or personal identifiable information (PII)}.
However, depending your applicable context and policies, you may want to provide this argument as a secret.}

\item{resource_group}{The \href{https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/manage-resource-groups-portal}{Azure resource group} to which the shiny app should belong.}

\item{docker_registry_server_url}{The container registry server url.}

\item{docker_registry_server_user}{The container registry server username.}
}
\description{
Wraps the \href{https://docs.microsoft.com/en-us/cli/azure/?view=azure-cli-latest}{Azure Command-Line Interface (CLI)} with defaults suitable for deploying a shiny app.
}
